requirements.txt
streamlit
langchain
openai
python-dotenv
python-docx
langchain-community